feat: split renderers, properly update size using client size
